Integrated ObfuscationMap into Sanitizer.
So process is now:

1) Add code to ObfuscationMap to identify obfuscated types, and
associate a new class name for them/

2) Sanitizer will then use these names when deobfuscating.

3) written classes will now be more legible.

Note the implementation atm isn't great; obfuscated types inside
obfuscated packages won't work properly.
Also it's possible for the name given in the ObfuscationMap to trigger
the deobfuscator logic. (if the meaningful name chosen in ObfuscationMap
has something like '0000')