Put in a fix for deobfuscation of static method invocations.
In .8 there was a static method invocation targeting a class, but the
static method itself was declared in a super class.

Didn't realize this was legal bytecode.
When searching for a deobfuscation name, it now checks super classes if
it initially fails to find a name.
It's not a perfect fix (it will easily mask bad bytecode).... but it
fixes this particular issue.