New Approach for 'installation'
Will replace starfarer_obj.jar with a dummy jar containing a dummy
StarfarerLauncher.

This dummy Launcher will attach an agent to the VM, add the proper
starfarer_obj.jar and SSME dependencies to the classpath, and pass
control through to the proper StarfarerLauncher.

So no matter how Starsector is loaded, the interception will succeed,
and there'll be no need to mess with commandline parameters. (hopefully
there'll be no increased memory footprint requirements)

This commit contains changes to the build script so it now builds the
assembly.
Needs further work, so it actually deploys it to the starsector mods
folder for integration testing.